Add link to best-board.games
We're trying to increase traffic on all of our sites, let's see if this
helps. For a start I chose to do the most basic version in the least
risky place (at the bottom of the page).

Next iterations could include placing the links at the top of the page
or below the fold and refining the link/ad itself.